Rename package maquina -> machine, and couple vars for consistency
