+++ machine/classic/common.go
