+++ machine/classic/contract.go
