+++ machine/classic/instructions.go
