+++ machine/classic/memory.go
