+++ machine/classic/stack.go
